refactor: issue-related subservices into dedicated files
