+++ issue.go
@@ -208,199 +208,6 @@ func (s *IssueService) Participants(ctx context.Context, issueIDOrKey string) ([
 }
 
 // ──────────────────────────────────────────────────────────────
-//  IssueAttachmentService
-// ──────────────────────────────────────────────────────────────
-
-// IssueAttachmentService handles communication with the issue attachment-related methods of the Backlog API.
-type IssueAttachmentService struct {
-	base *issue.AttachmentService
-}
-
-// List returns a list of all attachments in the issue.
-//
-// Backlog API docs: https://developer.nulab.com/docs/backlog/api/2/get-list-of-issue-attachments
-func (s *IssueAttachmentService) List(ctx context.Context, issueIDOrKey string) ([]*Attachment, error) {
-	v, err := s.base.List(ctx, issueIDOrKey)
-	return attachmentsFromModel(v), convertError(err)
-}
-
-// Remove removes a file attached to the issue.
-//
-// Backlog API docs: https://developer.nulab.com/docs/backlog/api/2/delete-issue-attachment
-func (s *IssueAttachmentService) Remove(ctx context.Context, issueIDOrKey string, attachmentID int) (*Attachment, error) {
-	v, err := s.base.Remove(ctx, issueIDOrKey, attachmentID)
-	return attachmentFromModel(v), convertError(err)
-}
-
-// Download downloads a file attached to the issue.
-// The caller is responsible for closing FileData.Body after use.
-//
-// Backlog API docs: https://developer.nulab.com/docs/backlog/api/2/get-issue-attachment
-func (s *IssueAttachmentService) Download(ctx context.Context, issueIDOrKey string, attachmentID int) (*FileData, error) {
-	v, err := s.base.Download(ctx, issueIDOrKey, attachmentID)
-	return fileDataFromModel(v), convertError(err)
-}
-
-// ──────────────────────────────────────────────────────────────
-//  IssueCommentService
-// ──────────────────────────────────────────────────────────────
-
-// IssueCommentService handles communication with the issue comment-related methods of the Backlog API.
-type IssueCommentService struct {
-	base   *issue.CommentService
-	Option *IssueCommentOptionService
-}
-
-// All returns a list of comments on an issue.
-//
-// This method supports options returned by methods in "*Client.Issue.Comment.Option",
-// such as:
-//   - WithMinID
-//   - WithMaxID
-//   - WithCount
-//   - WithOrder
-//
-// Backlog API docs: https://developer.nulab.com/docs/backlog/api/2/get-comment-list
-func (s *IssueCommentService) All(ctx context.Context, issueIDOrKey string, opts ...RequestOption) ([]*Comment, error) {
-	v, err := s.base.All(ctx, issueIDOrKey, toCoreOptions(opts)...)
-	return commentsFromModel(v), convertError(err)
-}
-
-// Add adds a comment to an issue.
-//
-// This method supports options returned by methods in "*Client.Issue.Comment.Option",
-// such as:
-//   - WithNotifiedUserIDs
-//   - WithAttachmentIDs
-//
-// Backlog API docs: https://developer.nulab.com/docs/backlog/api/2/add-comment
-func (s *IssueCommentService) Add(ctx context.Context, issueIDOrKey string, content string, opts ...RequestOption) (*Comment, error) {
-	v, err := s.base.Add(ctx, issueIDOrKey, content, toCoreOptions(opts)...)
-	return commentFromModel(v), convertError(err)
-}
-
-// Count returns the number of comments on an issue.
-//
-// Backlog API docs: https://developer.nulab.com/docs/backlog/api/2/count-comment
-func (s *IssueCommentService) Count(ctx context.Context, issueIDOrKey string) (int, error) {
-	count, err := s.base.Count(ctx, issueIDOrKey)
-	return count, convertError(err)
-}
-
-// One returns information about a specific comment.
-//
-// Backlog API docs: https://developer.nulab.com/docs/backlog/api/2/get-comment
-func (s *IssueCommentService) One(ctx context.Context, issueIDOrKey string, commentID int) (*Comment, error) {
-	v, err := s.base.One(ctx, issueIDOrKey, commentID)
-	return commentFromModel(v), convertError(err)
-}
-
-// Delete deletes a comment from an issue.
-//
-// Backlog API docs: https://developer.nulab.com/docs/backlog/api/2/delete-comment
-func (s *IssueCommentService) Delete(ctx context.Context, issueIDOrKey string, commentID int) (*Comment, error) {
-	v, err := s.base.Delete(ctx, issueIDOrKey, commentID)
-	return commentFromModel(v), convertError(err)
-}
-
-// Update updates a comment on an issue.
-//
-// Backlog API docs: https://developer.nulab.com/docs/backlog/api/2/update-comment
-func (s *IssueCommentService) Update(ctx context.Context, issueIDOrKey string, commentID int, content string) (*Comment, error) {
-	v, err := s.base.Update(ctx, issueIDOrKey, commentID, content)
-	return commentFromModel(v), convertError(err)
-}
-
-// Notifications returns a list of notifications on a comment.
-//
-// Backlog API docs: https://developer.nulab.com/docs/backlog/api/2/get-list-of-comment-notifications
-func (s *IssueCommentService) Notifications(ctx context.Context, issueIDOrKey string, commentID int) ([]*Notification, error) {
-	v, err := s.base.Notifications(ctx, issueIDOrKey, commentID)
-	return notificationsFromModel(v), convertError(err)
-}
-
-// Notify sends notifications for a comment.
-//
-// Backlog API docs: https://developer.nulab.com/docs/backlog/api/2/add-comment-notification
-func (s *IssueCommentService) Notify(ctx context.Context, issueIDOrKey string, commentID int, userIDs []int) (*Comment, error) {
-	v, err := s.base.Notify(ctx, issueIDOrKey, commentID, userIDs)
-	return commentFromModel(v), convertError(err)
-}
-
-// ──────────────────────────────────────────────────────────────
-//  IssueCommentOptionService
-// ──────────────────────────────────────────────────────────────
-
-// IssueCommentOptionService provides a domain-specific set of option builders
-// for operations within the IssueCommentService.
-type IssueCommentOptionService struct {
-	base *core.OptionService
-}
-
-// WithCount sets the number of comments to retrieve (1-100).
-func (s *IssueCommentOptionService) WithCount(count int) RequestOption {
-	return s.base.WithCount(count)
-}
-
-// WithMaxID filters comments with ID at or below the given value.
-func (s *IssueCommentOptionService) WithMaxID(id int) RequestOption {
-	return s.base.WithMaxID(id)
-}
-
-// WithMinID filters comments with ID at or above the given value.
-func (s *IssueCommentOptionService) WithMinID(id int) RequestOption {
-	return s.base.WithMinID(id)
-}
-
-// WithOrder sets the sort order of results.
-func (s *IssueCommentOptionService) WithOrder(order Order) RequestOption {
-	return s.base.WithOrder(string(order))
-}
-
-// WithNotifiedUserIDs returns an option to set multiple `notifiedUserId[]` parameters.
-func (s *IssueCommentOptionService) WithNotifiedUserIDs(ids []int) RequestOption {
-	return s.base.WithNotifiedUserIDs(ids)
-}
-
-// WithAttachmentIDs returns an option to set multiple `attachmentId[]` parameters.
-func (s *IssueCommentOptionService) WithAttachmentIDs(ids []int) RequestOption {
-	return s.base.WithAttachmentIDs(ids)
-}
-
-// ──────────────────────────────────────────────────────────────
-//  IssueSharedFileService
-// ──────────────────────────────────────────────────────────────
-
-// IssueSharedFileService handles communication with the issue shared-file-related methods of the Backlog API.
-type IssueSharedFileService struct {
-	base *issue.SharedFileService
-}
-
-// List returns a list of shared files linked to the issue.
-//
-// Backlog API docs: https://developer.nulab.com/docs/backlog/api/2/get-list-of-linked-shared-files
-func (s *IssueSharedFileService) List(ctx context.Context, issueIDOrKey string) ([]*SharedFile, error) {
-	v, err := s.base.List(ctx, issueIDOrKey)
-	return sharedFilesFromModel(v), convertError(err)
-}
-
-// Link links shared files to the issue.
-//
-// Backlog API docs: https://developer.nulab.com/docs/backlog/api/2/link-shared-files-to-issue
-func (s *IssueSharedFileService) Link(ctx context.Context, issueIDOrKey string, fileIDs []int) ([]*SharedFile, error) {
-	v, err := s.base.Link(ctx, issueIDOrKey, fileIDs)
-	return sharedFilesFromModel(v), convertError(err)
-}
-
-// Unlink removes a shared file link from the issue.
-//
-// Backlog API docs: https://developer.nulab.com/docs/backlog/api/2/remove-link-to-shared-file-from-issue
-func (s *IssueSharedFileService) Unlink(ctx context.Context, issueIDOrKey string, fileID int) (*SharedFile, error) {
-	v, err := s.base.Unlink(ctx, issueIDOrKey, fileID)
-	return sharedFileFromModel(v), convertError(err)
-}
-
-// ──────────────────────────────────────────────────────────────
 //  IssueStarService
 // ──────────────────────────────────────────────────────────────
 
@@ -674,25 +481,6 @@ func newIssueService(method *core.Method, option *core.OptionService) *IssueServ
 		Star:       newIssueStarService(method, option),
 
 		Option: newIssueOptionService(option),
-	}
-}
-
-func newIssueAttachmentService(method *core.Method) *IssueAttachmentService {
-	return &IssueAttachmentService{
-		base: issue.NewAttachmentService(method),
-	}
-}
-
-func newIssueCommentService(method *core.Method, option *core.OptionService) *IssueCommentService {
-	return &IssueCommentService{
-		base:   issue.NewCommentService(method),
-		Option: &IssueCommentOptionService{base: option},
-	}
-}
-
-func newIssueSharedFileService(method *core.Method) *IssueSharedFileService {
-	return &IssueSharedFileService{
-		base: issue.NewSharedFileService(method),
 	}
 }
